requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Flask-Bcrypt==1.0.1
mysql-connector-python==8.4.0
python-dotenv==1.0.1
requests==2.32.3
pytz==2024.1
